+++ src/strings/mod.rs
@@ -6,4 +6,4 @@ pub use formatting::{
     DecoratedString, StringDecoration, StringDecorator, StringFormatting, TerminalString,
 };
 pub use grid_string::GridStringIterator;
-pub use parse::{TerminalStringToken, parse_tokens};
+//pub use parse::{TerminalStringToken, parse_tokens};